Release 0.1

+++ ZLocation/ZLocation.Storage.psm1
@@ -1,14 +1,48 @@
-﻿Set-StrictMode -Version Latest
+Set-StrictMode -Version Latest
+
+$script:alreadyFailed = $false
+
+function Get-ZLocationBackupFilePath
+{
+    return (Join-Path $env:HOMEDRIVE (Join-Path $env:HOMEPATH 'z-location.txt'))
+}
+
+function Get-ZLocationPipename
+{
+    return 'zlocation'
+}
+
+#
+# Return cached proxy, or create a new one, if -Force
+#
+function Get-ZServiceProxy
+{
+    param(
+        [switch]$Force
+    )
+
+    if ((-not (Test-Path variable:Script:pipeProxy)) -or $Force) 
+    {
+        Set-Types
+        $pipeFactory = New-Object -TypeName 'System.ServiceModel.ChannelFactory`1[[ZLocation.IService]]' -ArgumentList @(
+            (Get-Binding),        
+            ( New-Object -TypeName 'System.ServiceModel.EndpointAddress' -ArgumentList ( $baseAddress + '/' + (Get-ZLocationPipename) ) )
+        )    
+        $Script:pipeProxy = $pipeFactory.CreateChannel()
+    }
+    $Script:pipeProxy
+}
 
 #
 # Return ready-to-use ZLocation.IService proxy.
 # Starts service server side, if nessesary
+# There is an issue https://github.com/vors/ZLocation/issues/1
+# We still cannot garante 100% availability.
+# We want to fail gracefully, and print warning.
 #
 function Get-ZService()
 {
     $baseAddress = "net.pipe://localhost"
-    $pipename = "zlocation"
-    $backupFilePath = Join-Path $env:HOMEDRIVE (Join-Path $env:HOMEPATH "z-location.txt")
 
     function log([string] $message)
     {
@@ -42,7 +76,7 @@ function Get-ZService()
     {
         if (-not (Test-Path variable:Script:binding)) {
             log "Create new .NET pipe service binding"
-            $Script:binding = [System.ServiceModel.NetNamedPipeBinding]::new()
+            $Script:binding = New-Object -TypeName 'System.ServiceModel.NetNamedPipeBinding'
             $Script:binding.OpenTimeout = [timespan]::MaxValue
             $Script:binding.CloseTimeout = [timespan]::MaxValue
             $Script:binding.ReceiveTimeout = [timespan]::MaxValue
@@ -52,28 +86,15 @@ function Get-ZService()
     }
 
     #
-    # Return cached proxy, or create a new one, if -Force
-    #
-    function Get-ZServiceProxy([switch]$Force)
-    {
-        if ((-not (Test-Path variable:Script:pipeProxy)) -or $Force) 
-        {
-            Set-Types
-            $pipeFactory = [System.ServiceModel.ChannelFactory[ZLocation.IService]]::new(
-                (Get-Binding), 
-                [System.ServiceModel.EndpointAddress]::new("$($baseAddress)/$($pipename)"))    
-            $Script:pipeProxy = $pipeFactory.CreateChannel()
-        }
-        $Script:pipeProxy
-    }
-
-    #
     # 
     #
     function Start-ZService()
     {
         Set-Types
-        $service = [System.ServiceModel.ServiceHost]::new([ZLocation.Service]::new($backupFilePath), [uri]($baseAddress))
+        $service = New-Object 'System.ServiceModel.ServiceHost' -ArgumentList (
+            (New-Object 'ZLocation.Service' -ArgumentList @( (Get-ZLocationBackupFilePath) ) ), 
+            [uri]($baseAddress)
+        )
 
         # It will be usefull to add debugBehaviour, like this
         # $debugBehaviour = $service.Description.Behaviors.Find[System.ServiceModel.Description.ServiceDebugBehavior]();
@@ -81,12 +102,14 @@ function Get-ZService()
         # $debugBehaviour.IncludeExceptionDetailInFaults = $true
         # $service.Description.Behaviors.Add($debugBehaviour);
 
-        $service.AddServiceEndpoint([ZLocation.IService], (Get-Binding), $pipename) > $null
+        $service.AddServiceEndpoint([ZLocation.IService], (Get-Binding), (Get-ZLocationPipename) ) > $null
         $service.Open() > $null
     }
 
     $service = Get-ZServiceProxy
     $retryCount = 0
+    
+    # This while loop is horible, sorry future me.
     while ($true) 
     {
         $retryCount++
@@ -97,35 +120,70 @@ function Get-ZService()
             if ($retryCount -gt 1)
             {
                 Write-Error "Cannot connect to a storage service. $_"
-                break;
+                return $null;
             }
-            Start-ZService
-            $service = Get-ZServiceProxy -Force
+            try
+            {
+                Start-ZService
+                $service = Get-ZServiceProxy -Force
+            } catch {
+                # This is the codepath that cause rear problems with broken pipe (https://github.com/vors/ZLocation/issues/1)
+                return $null
+            }
+            
         }
     }
 
-    $service
+    return $service
+}
+
+function Fail-Gracefully
+{
+    if (-not $script:alreadyFailed) {
+        Write-Warning @'
+ZLocation Pipe become broken :( ZLocation is now self-disabled. 
+You need to restart all PowerShell instances to re-enable ZLocation.
+Please continue your work and do it, when convinient.
+You can report the problem on https://github.com/vors/ZLocation/issues
+'@
+        $script:alreadyFailed = $true
+    }
 }
 
 function Get-ZLocation()
 {
     $service = Get-ZService
     $hash = @{}
-    foreach ($item in $service.Get()) 
+    if ($service) 
     {
-        $hash.add($item.Key, $item.Value)
-    }    
+        foreach ($item in $service.Get()) 
+        {
+            $hash.add($item.Key, $item.Value)
+        }    
+    } else {
+        Fail-Gracefully
+    }
     return $hash
 }
 
 function Add-ZWeight([string]$path, [double]$weight) {
     $service = Get-ZService
-    $service.Add($path, $weight)
+    if ($service)
+    {
+        $service.Add($path, $weight)
+    } else {
+        Fail-Gracefully    
+    }
 }
 
 function Remove-ZLocation([string]$path) {
     $service = Get-ZService
-    $service.Remove($path)
+    if ($service) 
+    {
+        $service.Remove($path)
+    } else {
+        Fail-Gracefully
+    }
 }
 
 $MyInvocation.MyCommand.ScriptBlock.Module.OnRemove = {
